+++ CHANGELOG.md
@@ -11,6 +11,8 @@ of [keepachangelog.com][2].
 
 ### Added
 
+- [#21](https://github.com/jaredbeck/minitest_to_rspec/pull/21) -
+  convert `setup`/`teardown` methods to `before`/`after` blocks
 - [#18](https://github.com/jaredbeck/minitest_to_rspec/pull/18) -
   Support namespaced exceptions for assert_raise[s]